圣诞快乐 # v1.3 更新列表     1. 【新增】砍价 	2. 【新增】拼团 	3. 【新增】一号通 	4. 【修复】商品sku 编辑时出现商品属性对应错误的问题 	5. 【修复】商品推广海报生成二维码可能会出错的问题【小程序调试中】 	6. 【修复】微信公众号和小程序头像可能获取不到的问题 	7. 【修复】下单时可能会出错的问题 	8. 【修复】pc管理端用户访问量 	9. 【修复】微信退款 	10. 【修复】管理端订单状态可能出现不正确的情况 	11. 【修复】WEB管理端-菜单色调，短信API更新，首页用户访问量，系统设置tab是自动选择下一及表单 	12. 【修复】系统设置出现更新不正确的问题

+++ crmeb/src/main/java/com/zbkj/crmeb/front/response/StoreOrderDetailResponse.java
@@ -130,7 +130,7 @@ public class StoreOrderDetailResponse {
     private BigDecimal gainIntegral;
 
     @ApiModelProperty(value = "使用积分")
-    private BigDecimal useIntegral;
+    private Integer useIntegral;
 
     @ApiModelProperty(value = "给用户退了多少积分")
     private BigDecimal backIntegral;
